Rename kinto-client to kinto-http.
