+++ kinto_http/cli_utils.py
