+++ kinto_http/exceptions.py
