+++ kinto_http/tests/config/kinto.ini
