+++ kinto_http/utils.py
