+++ tox.ini
@@ -4,11 +4,11 @@ envlist = py27,py34,flake8
 [testenv]
 commands =
     python --version
-    py.test kinto_client/tests/functional.py kinto_client/tests --cov-report term-missing --cov-fail-under 100 --cov kinto_client {posargs}
+    py.test kinto_http/tests/functional.py kinto_http/tests --cov-report term-missing --cov-fail-under 100 --cov kinto_http {posargs}
 deps = -rdev-requirements.txt
 install_command = pip install --pre {opts} {packages}
 
 [testenv:flake8]
-commands = flake8 kinto_client
+commands = flake8 kinto_http
 deps =
     flake8
